Update letter model, remove translate.py
